@@ -2,13 +2,13 @@ apply plugin: 'com.android.application'
 apply plugin: 'org.greenrobot.greendao'
 
 android {
-    compileSdkVersion 33
+    compileSdk 36
     defaultConfig {
         applicationId "dev.ukanth.iconmanager"
         minSdkVersion 21
-        targetSdkVersion 33
-        versionCode 20220820
-        versionName "4.0.0"
+        targetSdk 36
+        versionCode 20251230
+        versionName "5.0.0"
         testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
 
@@ -31,28 +31,25 @@ greendao {
 
 
 dependencies {
-    def libsuVersion = '5.0.2'
-  /*  compile fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-*/
-    implementation 'androidx.appcompat:appcompat:1.6.1'
+    def libsuVersion = '6.0.0'
+
+    implementation 'androidx.appcompat:appcompat:1.7.1'
     implementation 'androidx.cardview:cardview:1.0.0'
-    implementation 'androidx.recyclerview:recyclerview:1.3.0'
+    implementation 'androidx.recyclerview:recyclerview:1.4.0'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'androidx.legacy:legacy-support-v13:1.0.0'
-    implementation 'com.google.android.material:material:1.6.0'
-    implementation 'androidx.annotation:annotation:1.5.0'
-    implementation 'androidx.vectordrawable:vectordrawable:1.1.0'
-    implementation 'androidx.vectordrawable:vectordrawable-animated:1.1.0'
-    implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.1.0'
-    implementation "androidx.constraintlayout:constraintlayout:2.1.4"
+    implementation 'com.google.android.material:material:1.13.0'
+    implementation 'androidx.annotation:annotation:1.9.1'
+    implementation 'androidx.vectordrawable:vectordrawable:1.2.0'
+    implementation 'androidx.vectordrawable:vectordrawable-animated:1.2.0'
+    implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.2.0'
+    implementation "androidx.constraintlayout:constraintlayout:2.2.1"
 
     implementation 'androidx.palette:palette:1.0.0'
 
     //root library
     //implementation 'eu.chainfire:libsuperuser:1.0.0.201602271131'
+    //noinspection GradleDependency
     implementation "com.github.topjohnwu.libsu:core:${libsuVersion}"
 
     //database
@@ -61,20 +58,21 @@ dependencies {
     //material dialog
     implementation 'com.afollestad.material-dialogs:core:0.9.6.0'
 
-    //changelib
-    implementation 'com.github.gabrielemariotti.changeloglib:changelog:2.1.0'
-
     //gson
-    implementation 'com.google.code.gson:gson:2.9.1'
+    implementation 'com.google.code.gson:gson:2.11.0'
 
     //prettytime
     implementation 'org.ocpsoft.prettytime:prettytime:5.0.4.Final'
 
     //room
-    def room_version = "2.5.2"
+    def room_version = "2.6.1"
 
+    //noinspection GradleDependency
     implementation "androidx.room:room-runtime:$room_version"
     annotationProcessor "androidx.room:room-compiler:$room_version"
 
-
+    // Testing
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.2.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.6.1'
 }
